+++ sdks/paas-service/paas_service/constants.py
@@ -18,12 +18,23 @@ to the current version of the project delivered to anyone in the future.
 """
 from enum import Enum
 
+from blue_krill.data_types.enum import StrStructuredEnum, EnumField
+
 
 class Category(int, Enum):
     """Paas service categories"""
 
     DATA_STORAGE = 1
     MONITORING_HEALTHY = 2
+
+
+class ProvisionRecordStatus(StrStructuredEnum):
+    # 实例创建中
+    PROVISIONING = EnumField("provisioning", label="分配资源中")
+    # 成功：物理资源已创建
+    SUCCESS = EnumField("success", label="分配成功")
+
+    # 异步删除触发，或分配发生错误，都会直接删除 ProvisionRecord，所以无需对应的状态
 
 
 # Login 服务的重定向链接字段名